refresh: add subcommand to refresh all contract details

+++ uaclient/config.py
@@ -36,6 +36,7 @@ class UAConfig(object):
         'machine-access-support': 'machine-access-support.json',
         'machine-detach': 'machine-detach.json',
         'machine-token': 'machine-token.json',
+        'machine-token-refresh': 'machine-token-refresh.json',
         'macaroon': 'sso-macaroon.json',
         'root-macaroon': 'root-macaroon.json',
         'oauth': 'sso-oauth.json'
@@ -161,16 +162,26 @@ class UAConfig(object):
             return os.path.join(data_dir, self.data_paths[key])
         return os.path.join(data_dir, key)
 
+    def delete_cache_key(self, key):
+        """Remove specific cache file."""
+        if not key:
+            raise RuntimeError(
+                'Invalid or empty key provided to delete_cache_key')
+        if key.startswith('machine-access'):
+            self._entitlements = None
+        elif key == 'account-contracts':
+            self._contracts = None
+        elif key == 'machine-token':
+            self._machine_token = None
+        for private in (True, False):
+            cache_path = self.data_path(key, private)
+            if os.path.exists(cache_path):
+                os.unlink(cache_path)
+
     def delete_cache(self):
-        """Remove all configuration cached response files class attributes."""
-        self._contracts = None
-        self._entitlements = None
-        self._machine_token = None
-        for key in self.data_paths.keys():
-            for private in (True, False):
-                cache_path = self.data_path(key, private)
-                if os.path.exists(cache_path):
-                    os.unlink(cache_path)
+        """Remove configuration cached response files class attributes."""
+        for path_key in self.data_paths.keys():
+            self.delete_cache_key(path_key)
 
     def read_cache(self, key, quiet=False):
         cache_path = self.data_path(key)
